added mlx functions to cub3d.c; new makefile for ubuntu
