+++ alloc_check.c
@@ -6,7 +6,7 @@
 /*   By: abuzdin <abuzdin@student.s19.be>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/20 09:29:45 by abuzdin           #+#    #+#             */
-/*   Updated: 2022/08/15 14:28:20 by abuzdin          ###   ########.fr       */
+/*   Updated: 2022/08/16 13:01:17 by abuzdin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,6 +19,7 @@ void	alloc_check_big(char **str, t_input *data)
 	{
 		write(2, "Cub3d: ", 7);
 		write(2, "allocation failed\n", 18);
+		cub_free_all(data);
 		exit(1);
 	}
 }
@@ -29,6 +30,7 @@ void	alloc_check_small(void *str, t_input *data)
 	{
 		write(2, "Cub3d: ", 7);
 		write(2, "allocation failed\n", 18);
+		cub_free_all(data);
 		exit(1);
 	}
 }
