+++ cub3d.h
@@ -54,22 +54,25 @@ typedef struct s_input
 void	alloc_check_big(char **str, t_input *data);
 void	alloc_check_small(void *str, t_input *data);
 
-//	utils
+//	utils_1
 char	*cub_strjoin_free(char *rest, char *buf, t_input *data);
 char	*cub_charjoin_free(char *line, char b, t_input *data);
 char	*cub_strndup(char const *str, size_t size, t_input *data);
 char	*cub_strdup(const char *s, t_input *data);
 void	*cub_malloc(size_t n, t_input *data);
+
+//	utils_2
 int		ft_strcmp(char *s1, char *s2);
 void	my_mlx_pixel_put(t_img *img, int x, int y, int color);
 char	*get_next_line(int fd, t_input *data);
 int		check_charset(char c, char *charset);
+int		ft_exit(t_input *data);
 
 //	error_messages
 int		error_check_exit(int input, char *str, t_input *data);
 int		error_check_noexit(int input, char *str, t_input *data);
 void	error_exit(t_input *data, char *msg, int param);
-void	check_mlx(void *mlx);
+void	check_mlx(void *mlx, t_input *data);
 void	check_win(t_input *data);
 
 //	hooks
