+++ hooks.c
@@ -6,7 +6,7 @@
 /*   By: abuzdin <abuzdin@student.s19.be>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/29 11:55:14 by abuzdin           #+#    #+#             */
-/*   Updated: 2022/08/16 12:30:32 by abuzdin          ###   ########.fr       */
+/*   Updated: 2022/08/16 13:09:57 by abuzdin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,11 +14,12 @@
 
 int	key_hook(int keycode, t_input *data)
 {
-	if (keycode == 53)
+	if (keycode == 65307)
 	{
-		// mlx_destroy_image(data->mlx, data->img.mlx_img);
-		// mlx_destroy_window(data->mlx, data->win);
+		mlx_destroy_image(data->mlx, data->img.mlx_img);
+		mlx_destroy_window(data->mlx, data->win);
 		data->win = NULL;
+		cub_free_all(data);
 		exit(EXIT_SUCCESS);
 	}
 	// else if (keycode == 65361)
@@ -41,6 +42,10 @@ int	key_hook(int keycode, t_input *data)
 
 int	mouse_hook(int keycode, int x, int y, t_input *data)
 {
+	(void)x;
+	(void)y;
+	(void)keycode;
+	cub_free_all(data);
 	// if (keycode == 4)
 	// 	data->set.zoom *= 1.1;
 	// else if (keycode == 5)
